Consider wrongly paginated playlists
It seems that there are playlist which are wrongly paginated.
This means that if you ask for the second page of tracks you still get the same list of tracks as the first page consists of.
This was not considered because I only exited the request flow if the request contained no tracks at all.

Example for this kind of playlist is "37i9dQZF1CAl3F5fcqXZPo" "Your summer rewind".

By checking if all songs in the new request result are already present this case is considered too.