Avoid fetching same playlist multiple times
The useEffect is called multiple times.
There is the check that only those snapshotIds are requested which are not requested before.
But unfurtunately a snapshotId is registered only after the all of the requests were made.
Because this could take some time it results in a situation were the same playlist is requested multiple times anyway.
By directly registering those playlists which are going to be fetched before even fetching this problem can be avoided.