Use parallel requests for playlist artist ids query
Multiple queries for one playlist are delayed in the backend.
But I was frightend that if I send multiple parallel queries from the frontend to the backend that something bad would happen.
That spotify would give me TooManyRequests error. But this did not happend.

My guess would be that the individual serverless functions are counted separately from spotify or that the serverless functions
are by themselfe do not get executed at the same time.

But this solution yiels good results for now