Enable production deployment
This makes the component not any more staticly generated but dynamically.

`const { filter } = params.searchParams;` is only availble in client components.
And probably importing the lineup files by `await import(...)` too.

This could theoretically worked around or this simple workaround could be used.

It should be explorated if the file import is actually a problem in staticly generated pages.
And if it is somhow possible to put the searchParams usage in a child component to work around
making this component dynamic server component.